# docs/swagger.yaml
basePath: /
definitions:
  apierror.APIError:
    properties:
      code:
        type: integer
      sysMessage:
        type: string
      timestamp:
        type: string
      userMessage:
        type: string
    type: object
...
  domain.Receiver:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateReceiverRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        maxLength: 128
        type: string
    required:
    - email
    - name
    type: object
  handlers.CreateReceiverResponse:
    properties:
      id:
        example: c790g02f8n90184b23qg
        type: string
    type: object
...
  handlers.GetReceiverByIdResponse:
    properties:
      receiver:
        $ref: '#/definitions/domain.Receiver'
    type: object
...
  handlers.ListReceiversResponse:
    properties:
      receivers:
        items:
          $ref: '#/definitions/domain.Receiver'
        type: array
    type: object
...
  handlers.SendGroupRequest:
    properties:
      body:
        example: Important team updates...
        maxLength: 2048
        type: string
      group_id:
        example: c790g02f8n90184b23qg
        type: string
      subject:
        example: Team Announcement
        maxLength: 256
        type: string
    required:
    - body
    - group_id
    - subject
    type: object
  handlers.SendGroupWithTemplateRequest:
    properties:
      data: {}
      group_id:
        example: c790g02f8n90184b23qg
        type: string
      subject:
        example: Monthly Digest
        maxLength: 256
        type: string
      template_id:
        example: c790g02f8n90184b23qg
        type: string
    required:
    - data
    - group_id
    - subject
    - template_id
    type: object
  handlers.SendSingleRequest:
    properties:
      body:
        example: Hello, this is a plain text message.
        maxLength: 2048
        type: string
      subject:
        example: Welcome to Carrpigeon!
        maxLength: 256
        type: string
      to:
        example: receiver@example.com
        type: string
    required:
    - body
    - subject
    - to
    type: object
  handlers.SendSingleWithTemplateRequest:
    properties:
      data: {}
      subject:
        example: Your invoice
        maxLength: 256
        type: string
      template_id:
        example: c790g02f8n90184b23qg
        type: string
      to:
        example: receiver@example.com
        type: string
    required:
    - data
    - subject
    - template_id
    - to
    type: object
...
  handlers.UpdateReceiverRequest:
    properties:
      email:
        example: jane.doe@example.com
        type: string
      name:
        example: Jane Doe
        maxLength: 128
        type: string
    type: object
  handlers.UpdateReceiverResponse:
    properties:
      id:
        example: c790g02f8n90184b23qg
        type: string
    type: object
...
host: localhost:2500
info:
  contact: {}
  description: Carrpigeon email distribution and template service API.
  title: Carrpigeon API
  version: "1.0"
paths:
...
  /receivers:
    get:
      description: Get a paginated list of email receivers
      parameters:
      - default: 10
        description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListReceiversResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: List receivers
      tags:
      - receivers
    post:
      consumes:
      - application/json
      description: Creates a new email receiver
      parameters:
      - description: Receiver creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateReceiverRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.CreateReceiverResponse'
        "400":
          description: Invalid request or validation failed
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Create receiver
      tags:
      - receivers
  /receivers/{id}:
    delete:
      description: Deletes an email receiver by ID
      parameters:
      - description: Receiver ID (xid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Receiver Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Delete receiver
      tags:
      - receivers
    get:
      description: Get details of an email receiver by ID
      parameters:
      - description: Receiver ID (xid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetReceiverByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Receiver Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Get receiver by ID
      tags:
      - receivers
    put:
      consumes:
      - application/json
      description: Updates an existing receiver's name or email
      parameters:
      - description: Receiver ID (xid)
        in: path
        name: id
        required: true
        type: string
      - description: Receiver update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateReceiverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpdateReceiverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Receiver Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Update receiver
      tags:
      - receivers
  /send/group:
    post:
      consumes:
      - application/json
      description: Queues an email to all active receivers belonging to a specified
        group
      parameters:
      - description: Group email send payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendGroupRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request or validation error
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Send email to group
      tags:
      - send
  /send/group/template:
    post:
      consumes:
      - application/json
      description: Renders a template with provided data and queues an email to all
        members of the group
      parameters:
      - description: Group templated email send payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendGroupWithTemplateRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request or validation error
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group or template not found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Send email to group with template
      tags:
      - send
  /send/single:
    post:
      consumes:
      - application/json
      description: Queues a single plain text email to a specific recipient
      parameters:
      - description: Email send payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendSingleRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request or validation error
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Send single email
      tags:
      - send
  /send/single/template:
    post:
      consumes:
      - application/json
      description: Renders a template with provided data and queues an email to a
        single recipient
      parameters:
      - description: Templated email send payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendSingleWithTemplateRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request or validation error
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Send single email with template
      tags:
      - send
...
swagger: "2.0"